% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBindingStrength.R
\name{addBindingStrength}
\alias{addBindingStrength}
\title{Adds binding strength (essentailly same as GC content) to a bedfile.}
\usage{
addBindingStrength(bedFile, fastaFile, genome = "hg19", forceRedo = F)
}
\arguments{
\item{bedFile}{The bedfile.}

\item{fastaFile}{The fasta file.}
}
\description{
Adds binding strength (essentailly same as GC content) to a bedfile.
}
\details{
Adds the bidning strength as a column to the bedfile, saving it as a copy, replacing .bed with .dn.bed
}

