% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{analyse}
\alias{analyse}
\title{Analyse exomes}
\usage{
analyse(
  inputFiles,
  outputDirectories,
  settings,
  forceRedo,
  runtimeSettings,
  parameters = defaultSuperParameters(),
  byIndividual = T,
  manualStoryMerge = F,
  correctReferenceBias = T,
  vepCall = "vep",
  resourceDirectory = resourceDirectory,
  mode = "exome",
  filterOffTarget = T,
  rareGermline = T,
  annotationMethod = "VariantAnnotation",
  ploidyPriors = NULL
)
}
\arguments{
\item{inputFiles}{A named list of input files, containing the entries metaDataFile, vcfFiles, normalDirectory, captureRegionsFile and dbSNPdirectory}

\item{outputDirectories}{A named list of output directories, containing the entries Rdirectory and plotDirectory where the saved data and plots will be stored respectively.}

\item{settings}{A named list containing the entries genome and BQoffset. The only genome supporter atm is 'hg19', and the BQ offset is 33 for most exomes, altough some have 64. Check your fastqc files if you are not sure.}

\item{forceRedo}{A named list of logicals controling if existing saved data should be loaded or regenerated (overwriting the previous saved data). shortcuts to create these lists are forceRedoNothing() and forceRedoEverything().}

\item{runtimeSettings}{A named list containing the entries cpus and outputToTerminalAsWell. cpus is an integer controling the maximum number of cpus used in parallel, and outputToTerminalAsWell prints log data to the R session as well as to the log file.}
}
\description{
Analyse exomes
}
\details{
This function runs a full SNV, SNP, CNV and clonality analysis in the input exome data.
         Set up input data as is shown in the example.
}
\examples{
\dontrun{
metaDataFile = '/absolute/path/to/metaData.txt'

captureRegionsFile = '/absolute/path/to/captureRegions.bed'

dbSNPdirectory = '/absolute/path/to/dbSNP'

normalDirectory = '/path/to/normalDirectory'

Rdirectory = '/absolute/path/to/R'
plotDirectory = '/absolute/path/to/plots'

cpus=6

BQoffset = 33
genome = 'hg19'


inputFiles =
  list('metaDataFile'=metaDataFile, 'vcfFiles'=vcfFiles,
       'normalDirectory'=normalDirectory, 'normalCoverageDirectory'=normalCoverageDirectory,
       'captureRegionsFile'=captureRegionsFile, 'dbSNPdirectory'=dbSNPdirectory)
outputDirectories = list('Rdirectory'=Rdirectory, 'plotDirectory'=plotDirectory)

runtimeSettings = list('cpus'=cpus, 'outputToTerminalAsWell'=T)
settings = list('genome'=genome, 'BQoffset'=BQoffset)

forceRedo = forceRedoNothing()

parameters = list('systematicVariance'=0.02, 'maxCov'=150)

data = analyse(inputFiles, outputDirectories, settings, forceRedo, runtimeSettings, parameters)

}
}
\keyword{CNV}
\keyword{analyse}
\keyword{clonality}
\keyword{exomes}
