% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVariantAnnotation.R
\name{annotateQ}
\alias{annotateQ}
\title{Run VariantAnnotation on the provided variants}
\usage{
annotateQ(
  q,
  genome = "hg19",
  resourceDirectory = "superFreqResources",
  reference,
  cpus = 1
)
}
\arguments{
\item{q}{data.frame: The variants to be annotated.}

\item{genome}{character: hg19, hg38 or mm10.}

\item{reference}{character: Path to the reference fasta.}

\item{cpus}{integer: The number of cpus to be used as most.}

\item{annotationDirectory}{character: Where the annotation resources are taken from.}
}
\description{
Run VariantAnnotation on the provided variants
}
\details{
This function calls VEP on the output from outputSomaticVariants. For this, VEP needs to be callable by system('vep').
}
