% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVariantAnnotation.R
\name{annotateSomaticQs}
\alias{annotateSomaticQs}
\title{Run VariantAnnotation on the provided samples}
\usage{
annotateSomaticQs(
  qs,
  genome = "hg19",
  resourceDirectory = "superFreqResources",
  reference,
  cpus = 1
)
}
\arguments{
\item{qs}{list of data.frame: The variants to be annotated by sample}

\item{genome}{character: hg19, hg38 or mm10.}

\item{reference}{character: Path to the reference fasta.}

\item{cpus}{integer: The number of cpus to be used as most.}

\item{annotationDirectory}{character: Where the annotation resources are taken from.}
}
\description{
Run VariantAnnotation on the provided samples
}
\details{
This function uses the VariantAnnotation package to annotate somatic variants.
         PolyPHen, SIFT and exon number are removed when switching from VEP.
}
