% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToPileup.R
\name{bamToPileup}
\alias{bamToPileup}
\title{Gets pileup from bam over positions.}
\usage{
bamToPileup(
  bam,
  fasta,
  positions,
  index,
  Rdirectory,
  BQoffset = 33,
  samtoolsVersion = "samtools 1.9"
)
}
\description{
Gets pileup from bam over positions.
}
