% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCaptureRegions.R
\name{chrLengths}
\alias{chrLengths}
\title{The lengths of the chromsomes}
\usage{
chrLengths(genome = "hg19")
}
\arguments{
\item{genome}{character: the name of the genome. hg19 and mm10 supported atm.}
}
\description{
The lengths of the chromsomes
}
\details{
Returns a named vector of the chromosome lengths of the standard chromosomes.
}
\examples{
barplot(chrLengths('hg19'))
chrLengths('mm10')

#the chromosome boundaries in the genomic x coordinate.
plot(c(0, cumsum(chrLengths())), type='b')

}

