% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantsByIndividual.R
\name{chrToX}
\alias{chrToX}
\title{convert chromsome and position to the single genomic coordinate x}
\usage{
chrToX(chr, bp, genome = "hg19")
}
\arguments{
\item{chr}{character: The chromosome. Both 'chr1' and '1' works. 'Chr1' does not.}

\item{bp}{integer: the position on the chromosome.}

\item{genome}{character: the genome, such as 'hg19', 'hg38' or 'mm10'. Default 'hg19'.}
}
\description{
convert chromsome and position to the single genomic coordinate x
}
\details{
convert chromsome and position to the single genomic coordinate x.
}
\examples{
xToPos(1e9)
xToPos(1e9, genome='mm10')
}
