% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortAnalyseBatch.R
\name{cohortAnalyseBatch}
\alias{cohortAnalyseBatch}
\title{Analyse individuals for reccuring mutations}
\usage{
cohortAnalyseBatch(metaDataFile, outputDirectories, cpus = 1, onlyDNA = T,
  clonalityCut = 0.4, excludeSamples = c(), excludeIndividuals = c(),
  cosmicDirectory = "", analysisName = "cohortAnalysis",
  forceRedoVariants = F, forceRedoMean = F, forceRedoMatrixPlot = F,
  forceRedoMeanPlot = F, genome = "hg19")
}
\arguments{
\item{metaDataFile}{character: path to the metaData file.}

\item{outputDirectories}{A named list of output directories, containing the entries Rdirectory and plotDirectory where the saved data and plots will be stored respectively.}

\item{cpus}{integer: the maximum number of cpus to run on.}

\item{clonalityCut}{numeric: the minimum required clonality to be included in the analysis. Deafult 0.4.}

\item{excludeSamples}{character: The samples to be excluded from the analysis. Default c().}

\item{excludeIndividuals}{character: The individuals to be excluded from the analysis. Default c().}

\item{cosmicDirectory}{character: The directory with the COSMIC data.}

\item{analysisName}{character: The name of the directory where the analysis results are saved.
Default "cohortAnalysis".}

\item{forceRedoVariants}{boolean: Force redo the merging of the variants between individuals. Default FALSE.}

\item{forceRedoMean}{boolean: Force redo the mean CNAs ans SNVs rates over individuals. Default FALSE.}

\item{forceRedoMatrixPlot}{boolean: Force redo the hit matrix plot. Default FALSE.}

\item{forceRedoMeanPlot}{boolean: Force redo the mean CNA plot. Default FALSE.}

\item{genome}{character: the genome being studied. Default "hg19".}
}
\description{
Analyse individuals for reccuring mutations
}
\details{
This function calculates mutation rates over genes, both protein changing SNVs, as well as CNA rates for complete loss, loss, gain (3 copies) and amplification (4 or more copies). It also track biallelic loss of genes in samples, by complete loss, a protein changing SNV plus a loss, or two SNVs (that are assumed to be on different alleles). Output is a plot over the genome of the CNA rates, as well as a "top table" of frequently mutated genes. It is run as an afterburner, and needs a finished superFreq analysis to be run on the samples.
}
\examples{
\dontrun{
metaDataFile = '/absolute/path/to/metaData.txt'

Rdirectory = '/absolute/path/to/R'
plotDirectory = '/absolute/path/to/plots'

cpus=6
genome = 'hg19'

outputDirectories = list('Rdirectory'=Rdirectory, 'plotDirectory'=plotDirectory)

cohortAnalyseBatch(metaDataFile, outputDirectories, cpus=cpus, genome=genome)

}
}

