% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanCNV.R
\name{compareGroups}
\alias{compareGroups}
\title{Compares sets of individuals for reccuring mutations}
\usage{
compareGroups(metaDataFile, outputDirectories, project, subgroups1, subgroups2,
  name, clonalityCut = 0.4, excludeSamples = c(),
  excludeIndividuals = c(), cosmicDirectory = "",
  analysisName = "cohortAnalysis", cpus = 1, forceRedoVariants = F,
  forceRedoMean = F, forceRedoMeanPlot = F, forceRedoMatrixPlot = F,
  genome = "hg19")
}
\arguments{
\item{metaDataFile}{character: Path to the metaData.}

\item{outputDirectories}{A named list of output directories, containing the entry Rdirectory and plotDirectory where the saved data and plots will be stored respectively.}

\item{project}{character: The project containing the subgroups.}

\item{name}{character: The name of the comparison. This names the output directory.}

\item{clonalityCut}{numeric: the minimum required clonality to be included in the analysis. Deafult 0.4.}

\item{cosmicDirectory}{character: The directory with the COSMIC data.}

\item{cpus}{integer: The maximum number of parallel processes.}

\item{forceRedoMean}{boolean: Force redo the mean CNAs ans SNVs rates over individuals. Default FALSE.}

\item{forceRedoMeanPlot}{boolean: Force redo the mean CNA plot. Default FALSE.}

\item{forceRedoMatrixPlot}{boolean: Force redo the hit matrix plot. Default FALSE.}

\item{genome}{character: the genome being studied. Default "hg19".}

\item{subgroup1}{character: The first subgroup(s).}

\item{subgroup2}{character: The second subgroup(s).}
}
\description{
Compares sets of individuals for reccuring mutations
}
\details{
This function runs a cohort analysis, comparing two subgroups (within a project) to each other. See cohortAnalyseBatch for details.
}

