% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{defaultSuperParameters}
\alias{defaultSuperParameters}
\title{Returns default parameters}
\usage{
defaultSuperParameters(systematicVariance = "", maxCov = "",
  cloneDistanceCut = "")
}
\description{
Returns default parameters
}
\details{
feed the output of this function to analyse, or just call superFreq.
}
\examples{
defaultSuperParameters()
}

