% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{defaultSuperRuntimeSettings}
\alias{defaultSuperRuntimeSettings}
\title{Returns default runtime settings}
\usage{
defaultSuperRuntimeSettings(cpus = "", outputToTerminalAsWell = "")
}
\description{
Returns default runtime settings
}
\details{
feed the output of this function to analyse, or just call superFreq.
}
\examples{
defaultSuperRuntimeSettings()
}
