% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{defaultSuperSettings}
\alias{defaultSuperSettings}
\title{Returns default settings}
\usage{
defaultSuperSettings(genome = "", BQoffset = "")
}
\description{
Returns default settings
}
\details{
feed the output of this function to analyse, or just call superFreq.
}
\examples{
defaultSuperSettings()
}

