% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{downloadTemplate}
\alias{downloadTemplate}
\title{Downloads resources and template for a standard analysis}
\usage{
downloadTemplate(analysisDirectory)
}
\arguments{
\item{analysisDirectory}{The directory where the analysis will be made. The directory will be created
if needed, but the parent directory has to exist.}
}
\description{
Downloads resources and template for a standard analysis
}
\details{
This function downloads dbSNP and COSMIC resources for hg19, condensed to what superFreq needs.
         It also sets up template files and directories to help set up the project specific data.
}
\examples{
\dontrun{
 
downloadTemplate('~/mySuperFreqAnalaysis')

}
}

