% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{dumpInput}
\alias{dumpInput}
\title{internal function for debugging purposes.}
\usage{
dumpInput(Rdirectory, inputList)
}
\description{
internal function for debugging purposes.
}
\details{
This function saves the input to a function in a directory, so that the exact input triggering the bug
         can be examined.
}

