% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{forceRedoEverything}
\alias{forceRedoEverything}
\title{returns input that uses saved data if present.}
\usage{
forceRedoEverything()
}
\description{
returns input that uses saved data if present.
}
\details{
This function returns the input 'forceRedo' for analyse(), so that all the
         analysis is redone, overwriting previously saved data exists.
}
\examples{
\dontrun{
cpus=4
metaDataFile = '/absolute/path/to/myAnalysis/metaData.txt'
vcfFiles = list.files('/absolute/path/to/myAnalysis/vcf', pattern='chr[0-9XYMT]+.vcf$', full.names=T)
captureRegionsFile = '/absolute/path/to/myAnalysis/captureRegions.gc.bed'
dbSNPdirectory = '/absolute/path/to/myAnalysis/dbSNP'
normalDirectory = '/absolute/path/to/myAnalysis/normal'
Rdirectory = '/absolute/path/to/myAnalysis/R/'
plotDirectory = '/absolute/path/to/myAnalysis/plots/'

#The base quality phred offset. This can be read from fastqc analysis for example.
BQoffset = 33
genome = 'hg19'

inputFiles = list('metaDataFile'=metaDataFile, 'vcfFiles'=vcfFiles, 'normalDirectory'=normalDirectory,
'captureRegionsFile'=captureRegionsFile, 'dbSNPdirectory'=dbSNPdirectory)
outputDirectories = list('Rdirectory'=Rdirectory, 'plotDirectory'=plotDirectory)
runtimeSettings = list('cpus'=cpus, 'outputToTerminalAsWell'=T)

forceRedo = forceRedoEverything()
parameters = list('systematicVariance'=0.02, 'maxCov'=150)

data = analyse(inputFiles, outputDirectories, settings, forceRedo, runtimeSettings, parameters=parameters)
}
}
