% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get104signature.R
\name{get104profile}
\alias{get104profile}
\title{returns the 104 signature from a superFreq q variant object}
\usage{
get104profile(q, genome, somaticPcut = 0.5)
}
\description{
returns the 104 signature from a superFreq q variant object
}
