% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVEP.R
\name{getCosmicCounts}
\alias{getCosmicCounts}
\title{Retrieves information about cosmic variant IDs}
\usage{
getCosmicCounts(cosmic,
  cosmicDirectory = "/wehisan/general/academic/grp_leukemia_genomics/data/resources/COSMIC",
  onlyCensus = T)
}
\arguments{
\item{cosmic}{character: The cosmic IDs.}

\item{cosmicDirectory}{character: The directory where the cosmic files CosmicMutantExportCensus.tsv and CosmicMutantExport.tsv are located.}

\item{onlyCensus}{logical: Restrict to cosmic census genes.}
}
\description{
Retrieves information about cosmic variant IDs
}
\details{
For each cosmic ID, return whether it is seen in a census gene, and the mutations per million tumors (MPM) for that variant, as well as the average MPM over the gene (mutations per million tumors per basepair, MPMPB).
}

