% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVEP.R
\name{getMoreVEPinfo}
\alias{getMoreVEPinfo}
\title{Import more information about the variants}
\usage{
getMoreVEPinfo(variants, plotDirectory, genome = "hg19",
  cosmicDirectory = "")
}
\arguments{
\item{variants}{variants: The variants}

\item{plotDirectory}{character: The plot directory from the analyse call.}
}
\description{
Import more information about the variants
}
\details{
Extract almost all the information from the VEP run, and cross-checks with COSMIC data as well.
         getCosmicCounts is called by this function, see details of that for requirements.
}

