% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{getSettings}
\alias{getSettings}
\title{Wrapper for getting settings, containing defaults for missing values.}
\usage{
getSettings(settings, name)
}
\arguments{
\item{settings}{A named list of settings.}

\item{name}{The setting to be retrieved.}
}
\description{
Wrapper for getting settings, containing defaults for missing values.
}
\details{
If the setting is present in the list, just return settings[[name]].
         Otherwise return a default value.
}
