% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStories.R
\name{getStories}
\alias{getStories}
\title{Combines variants and CNAs to clonal evolution}
\usage{
getStories(variants, normalVariants, cnvs, timeSeries, normals, genome,
  cloneDistanceCut = -qnorm(0.01), Rdirectory, plotDirectory, cpus = 1,
  forceRedo = F, manualStoryMerge = F)
}
\arguments{
\item{variants}{variants: The variants.}

\item{normalVariants}{variants: The normal variants.}

\item{cnvs}{cnvs: The copy number calls.}

\item{timeSeries}{names list of vectors: samples to be analysed together, named by individual.}

\item{normals}{named boolean vector: which sameples (names of vector) are normal.}

\item{genome}{character: the genome.}

\item{Rdirectory}{character: The save directory.}

\item{plotDirectory}{character: the directory to plot to.}

\item{cpus}{integer: maximum number of parallel processes. Default 1.}

\item{forceRedo}{boolen: if redoing calculations even if saved data is available. Default FALSE.}
}
\description{
Combines variants and CNAs to clonal evolution
}
\details{
This function calls VEP on the output from outputSomaticVariants. For this, VEP needs to be callable by system('vep').
}

