% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVariantsByIndividual.R
\name{getVariantsByIndividual}
\alias{getVariantsByIndividual}
\title{Quality controls the variants in the vcfs, using the bams.}
\usage{
getVariantsByIndividual(
  metaData,
  captureRegions,
  fasta,
  genome,
  BQoffset,
  dbDir,
  Rdirectory,
  plotDirectory,
  cpus,
  exacPopulation = "all",
  forceRedo = F,
  filterOffTarget = T
)
}
\description{
Quality controls the variants in the vcfs, using the bams.
}
