% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCaptureRegions.R
\name{importCaptureRegions}
\alias{importCaptureRegions}
\title{Imports capture regions from a bed file with GC information.}
\usage{
importCaptureRegions(
  bedFile,
  reference,
  Rdirectory,
  genome = "hg19",
  forceRedoCR = F,
  forceRedoRdirectorySave = F
)
}
\arguments{
\item{genome}{Which genome the bed file is on. 'hg19', 'hg38' and 'mm10' only supported atm.}

\item{file}{Bedfile of the capture regions with extra columns for gc and binding strength information.}

\item{gcColumn}{the column in the bed file that contains the gc content. Default 5 matches output from the provided script at X.}

\item{dnColumn}{the column in the bed file that contains the binding strength. Default 6 matches output from the provided script at X.}
}
\description{
Imports capture regions from a bed file with GC information.
}
\details{
Reads in the bed file into GRanges format, with extra columns for GC and binding strength.
}
\examples{
captureRegions = importCaptureRegions('path/to/my/captureRegions.bed')
plotColourScatter(captureRegions$gc, captureRegions$dn, xlab='GC content', ylab='binding strength')

}
