% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSampleMetaData.R
\name{importSampleMetaData}
\alias{importSampleMetaData}
\title{Imports metadata about the samples}
\usage{
importSampleMetaData(sampleMetaDataFile)
}
\description{
Imports metadata about the samples
}
\details{
This function imports metadata from a tab-separates file. The file should have a column name row, and then one row for each bam file. The metadata must include the following columns, but in any order.

         BAM  Relative path to the bamfile from the metadata file.
         NAME The name of the sample that will appear on plots and in output. Must be unique.
         INDIVIDUAL Unique identifier for the individual. Samples from the same individual will be compared to each other, mutations will be tracked over samples of the individual, and a normal sample from the same individual will be used as matched normal.
         NORMAL If the sample is normal (tumor burden at most 1%). 'YES' or 'NO'
}
\examples{
#import and print metadata
metaDataFile = '/absolute/path/to/myAnalysis/metaData.txt'
metaData = importMetaData(metaDataFile)
metaData

}

