% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{loadData}
\alias{loadData}
\title{Loads saved data}
\usage{
loadData(Rdirectory, setVariantLoss = F, correctReferenceBias = T)
}
\arguments{
\item{Rdirectory}{A character string pointing to the Rdirectory of the analysis.}

\item{setVariantLoss}{If the average variant loss should be calculated. This measures reference bias of the variant frequencies, is set globally and is used by many other functions.}
}
\description{
This function returns the data produced from an analyse() run.
}
\examples{
\dontrun{
not run
data = loadData('~/myAnalysis/R', setVariantLoss=T)
cnvs = data$clusters
plotCR(cnvs[['aSample']]$clusters)
}
}
\keyword{data}
\keyword{load}
\keyword{saved}
