% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{loadMethods}
\alias{loadMethods}
\title{Not needed in the package, ignore this function.}
\usage{
loadMethods(stringsAsFactors = FALSE, byIndividual = F)
}
\description{
This function loads the analysis functions used in analyse().
}
\examples{
loadMethods()
}
\keyword{analyse}
\keyword{load}
\keyword{methods}

