% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFitPlots.R
\name{makeFitPlots}
\alias{makeFitPlots}
\title{Plots results from differential coverage analysis}
\usage{
makeFitPlots(
  fit,
  plotDirectory,
  genome,
  forceRedoVolcanoes = F,
  forceRedoDifferentRegions = F
)
}
\description{
Plots results from differential coverage analysis
}
