% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRiverPlots.R
\name{makepHeatmap}
\alias{makepHeatmap}
\title{plots a heatmap}
\usage{
makepHeatmap(
  mx,
  nCol = 200,
  col = "default",
  maxVal = "default",
  minVal = "default",
  label = "",
  DEsaturation = log2(10),
  reverseGradient = F,
  border_color = NA,
  cluster_rows = F,
  cluster_cols = F,
  ...
)
}
\arguments{
\item{mx}{numeric. The matrix to be plotted.}

\item{nCol}{numeric. The number of colours in the gradient. Default 200.}

\item{col}{character. The colour scheme to be used. The default 'default' gives a scale adapted for freqwuencies between 0 and 1, with extra attention to values close to 0. 'sunset' is a easier on the eye, but has less ability to resolve values close to the lowest value. 'DE' is meant for differential expression, and sets 0 to black, with red and blue gradient towards positive and negative values, saturating at the DEsaturation parameter.}

\item{maxVal}{numeric. Only for internal use. Default 'default'.}

\item{minVal}{numeric. Only for internal use. Default 'default'.}

\item{label}{character. The label for the colour scale on the side. Default ''.}

\item{DEsaturation}{numeric. The value where the colour saturates when col='DE'. Default log2(10).}

\item{...}{remaining arguments are passed to base heatmap(...)}
}
\description{
plots a heatmap
}
\details{
This function is a wrapped pheatmap() function with some changed colours and defaults. Depends on pheatmap package, which is not a superFreq dependency, so need to load that library before using this.
}
\examples{
#random matrix to plot, centered around 0. Plot in 'DE' colours.
mx = matrix(rt(400, df=10), nrow=100)
makeHeatmap(mx, col='DE')

#random matrix to plot, between 0 and 1. Plot in default and sunset colours.
mx = matrix(runif(400), nrow=100)
makeHeatmap(mx)
makeHeatmap(mx, col='sunset')

}
