% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortAnalyseBatch.R
\name{mergeBatches}
\alias{mergeBatches}
\title{Merges data from several batches for cohort analysis}
\usage{
mergeBatches(paths, targetMetaDataFile, targetRdirectory)
}
\arguments{
\item{paths}{data.frame: One row for each batch to be merged.}

\item{targetMetaDataFile}{character: Path to where the merged metaData will be stored.
This also affect where a downstream cohort analysis is stored.}

\item{targetRdirectory}{character: Path to where the Rdata will be stored.}
}
\description{
Merges data from several batches for cohort analysis
}
\details{
This function merges the data from multiple batches, setting them up for a joint cohort analysis. Note that this is only for the purpose of a downstream cohort analysis. The output R directory cannot be run as a batch Rdirectory.
}
\examples{
\dontrun{
paths = data.frame('metaDataFile' = c('~/superFreq/myFirstBatch/metaData.txt', '~/superFreq/mySecondBatch/metaData.txt'),
           'Rdirectory'= c('~/superFreq/myFirstBatch/R', '~/superFreq/myFirstBatch/R'))

targetMetaDataFile = '~/superFreq/mergedAnalysis/metaData.txt'
targetRdirectory = '~/superFreq/mergedAnalysis/R'

mergeBatches(paths, targetMetaDataFile, targetRdirectory)

outputDirectories = list('Rdirectory'=targetRdirectory)
cohortAnalyseBatch(targetMetaDataFile, outputDirectories, cpus=6, genome='hg19')

}
}

