% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesOfInterest.R
\name{plotCNAheatmapOverGoI}
\alias{plotCNAheatmapOverGoI}
\title{plots a heatmap of copy numbers, focusing on a gene of interest}
\usage{
plotCNAheatmapOverGoI(gene, Rdirectory, genome, metaDataFile,
  excludeIndividuals = c(), excludeSamples = c(), GoIakas = c(),
  cpus = 1, padding = 3000000)
}
\arguments{
\item{Rdirectory}{character The Rdirectory where the superFreq run was run.}

\item{genome}{character. The genome the sample is aligned to. 'hg19', 'hg38' or 'mm10'}

\item{padding}{numeric. How many basepairs next to the gene to show. default 3e6.}

\item{GoI}{character vector. Genes of interest.}
}
\description{
plots a heatmap of copy numbers, focusing on a gene of interest
}
\details{
This function plots a heatmap of copy numbers across the samples, focused at a gene of interest. Can be run after the superFreq run is completed.
}
