% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesOfInterest.R
\name{plotCNAheatmapWithGoI}
\alias{plotCNAheatmapWithGoI}
\title{plots a heatmap of copy numbers, highlighting genes of interest}
\usage{
plotCNAheatmapWithGoI(
  GoI,
  Rdirectory,
  genome,
  metaDataFile,
  excludeIndividuals = c(),
  excludeSamples = c(),
  GoIakas = c(),
  cpus = 1
)
}
\arguments{
\item{GoI}{character vector. Genes of interest.}

\item{Rdirectory}{character The Rdirectory where the superFreq run was run.}

\item{genome}{character. The genome the sample is aligned to. 'hg19', 'hg38' or 'mm10'}
}
\description{
plots a heatmap of copy numbers, highlighting genes of interest
}
\details{
This function plots a heatmap of copy numbers across the samples and genome. Genes of Interest are highlighted. Can be run after the superFreq run is completed.
}
