% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesOfInterest.R
\name{plotCohortMutationHeatmap}
\alias{plotCohortMutationHeatmap}
\title{plots mutations in samples in Genes of Interest}
\usage{
plotCohortMutationHeatmap(
  GoI,
  Rdirectory,
  metaDataFile,
  excludeIndividuals = c(),
  excludeSamples = c(),
  GoIakas = c(),
  cpus = 1
)
}
\arguments{
\item{GoI}{character vector. Genes of interest for the analysis}

\item{Rdirectory}{character The Rdirectory where the superFreq run was run.}
}
\description{
plots mutations in samples in Genes of Interest
}
\details{
This function plots a mutation matrix showing CNAs and point mutations in the Genes of Interest across the analysed samples in the Rdirectory. Should be run after the superFreq() run has completed.
}
