% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCNVplots.R
\name{plotCR}
\alias{plotCR}
\title{plots a copy number profile}
\usage{
plotCR(cR, showClonality = T, errorBars = T, chr = "all",
  genome = "hg19", alpha = 1, add = F, moveHet = T, pt.cex = 1,
  setMargins = T, fullFrequency = F, colourDeviation = T,
  forceCol = NA, plotCall = T, plotArrows = F, smallPlot = F,
  lwd = 1, sep.lwd = 5, sideSpace = NULL, noSex = F, ...)
}
\arguments{
\item{cR}{data.frame. The "clusters" or "CR" from the data.}

\item{showClonality}{logical. if the clonality panel is shown, if data available. Default TRUE.}

\item{errorBars}{logical. If errorbars are plotted. Default TRUE.}

\item{chr}{character. Which chromosome to plot. Default 'all', which plots the entire genome.}

\item{genome}{character. The genome assembly. Default 'hg19'.}

\item{alpha}{numeric. The opacity of the point. Default 1.}

\item{add}{logical. If the data should be plotted on top of whatever is already there. Default FALSE.}

\item{moveHet}{logical. If the SNP data should snap to f=0.5 if not significantly different. Default TRUE.}

\item{pt.cex}{numerical. Scaling factor for the size of the points. Default 1.}

\item{setMargins}{logical. If the margins should be removed to use the entire plottable area. Default TRUE.}

\item{fullFrequency}{logical. If the SNPs should be copied back up to 1-f as well. Generally not a good idea, and potentially misleading, but can be useful for people that are not used to see frequencies mirrored down to 0-0.5.}

\item{colourDeviation}{logical. If the points should be coloured based on deviation from diploid. Defaulty TRUE.}

\item{forceCol}{colour. The colour all the points will be plotted in. Can be useful if overplotting several samples with add=T. Overrides colourDeviation if not NA. Default NA.}

\item{plotCall}{logical. If the copy number call should be added to the bottom of the SNP panel. Default TRUE.}

\item{...}{remaining arguments are passed to base plot(...) if add=F, otherwise ignored.}
}
\description{
plots a copy number profile
}
\details{
This function plots copy number information from superFreq, either on a by-gene level or by-segment.
}
