% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEexon.R
\name{plotMA}
\alias{plotMA}
\title{a plotting function for MA plots.}
\usage{
plotMA(x, y, col = mcri("darkblue"), libNorm = F, span = 0.2,
  medianSigma = T, xlab = "A = log2(x*y)/2", ylab = "M = log2(x/y)",
  loess = F, cex = 0.6, pch = 16, verbose = T, ...)
}
\arguments{
\item{x}{The x coordinates. Large x will push points towards up right.}

\item{y}{The y coordinates. Large y will push points towards up right.}

\item{col}{The base colour of the dots.}

\item{libNorm}{If library normalisation should be perform on x and y before plotting.}

\item{span}{The span of the loess fit. Higher values gives smoother fits.}

\item{medianSigma}{Boolean. Prints the median deviation from 0 compared to poisson standard deviation. This is a good measure of how much systematic differences there are between x and y.}

\item{loess}{Boolean. Plots a loess fit of the data.}

\item{verbose}{Boolean. Outputs some messages while plotting.}

\item{...}{Remaining arguments are passed on to plot(...).}
}
\description{
a plotting function for MA plots.
}
\details{
This plotting function does normal MA plots, with nice density visuals. Also have a range of options for things such as library normalisation and loess fits.
}
\examples{
x = rpois(3000, 1:3000)
y = rpois(3000, 1:3000)
plotMA(x, y)
plotMA(x, y, loess=T)

x = rpois(3000, round((1:3000)*1.2))
y = rpois(3000, 1:3000)
plotMA(x, y, loess=T)

}
