% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRiverPlots.R
\name{plotRiver}
\alias{plotRiver}
\title{plots a river plot}
\usage{
plotRiver(cloneTree, cloneStories, storyList, allStories, variants,
  genome = "hg19", normalise = T, xlim = "default", ylim = "default",
  labels = T, setPar = T, sampleOrder = "default", excludeClones = c(),
  markDodgy = T, colourPool = c(), ignoreStoriesBelowSigma = 2,
  smallPlot = F, annotationMethod = "VariantAnnotation",
  plotSampleNames = T, ...)
}
\arguments{
\item{variants}{dataFrame from superFreq. such as data$allVariants$variants$variants$mySample}

\item{genome}{character. The genome the sample is aligned to. 'hg19', 'hg38' or 'mm10'}

\item{xlim}{numerical. The X limits. Default "default' set it depending on number of samples and clones.}

\item{labels}{logical. If labels are plotted. Default TRUE.}

\item{setPar}{logical. If the margins are set. Default TRUE.}

\item{sampleOrder}{character. The order of the samples. Default NA retains the order in the input.}

\item{...}{remaining arguments are passed to base plot(...) if add=F, otherwise ignored.}

\item{stories}{dataFrame from superFreq. such as data$stories$stories$mySample$all.}

\item{col}{colour. The colour of the lines. Default 'default' sets different colours on all lines.}

\item{lty}{linetype. The line type of the lines. Default 'default' sets different line types on all lines.}

\item{lwd}{linetype. The line width of the lines. Default 'default' sets different line width from accuracy.}

\item{errorBars}{logical. If error bars are used. Default TRUE.}

\item{add}{logical. If the data should be plotted on top of whatever is already there. Default FALSE.}

\item{alpha}{numerical. A opaqueness multiplier. Default 1.}

\item{xlab}{character. The X label. Default "sample'}

\item{ylab}{character. The Y label. Default "clonality'}

\item{legend}{logical. If a legend is plotted. Default TRUE.}

\item{xSpread}{numerical. The stread of the points over a sample. Default 0.25.}
}
\description{
plots a river plot
}
\details{
This function plots the story lines from a superFreq analysis.
         Can be used both for individual mutations or for clones.
}
