% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRiverPlots.R
\name{plotStories}
\alias{plotStories}
\title{plots line plots of clonalities}
\usage{
plotStories(stories, variants, col = "default", lty = "default", add = F,
  alpha = 1, xlab = "sample", ylab = "clonality", lwd = "default",
  errorBars = T, setPar = T, legend = T, labels = T, xlim = "default",
  genome = "hg19", xSpread = 0.25, sampleOrder = NA,
  annotationMethod = "VariantAnnotation", ...)
}
\arguments{
\item{stories}{dataFrame from superFreq. such as data$stories$stories$mySample$all.}

\item{variants}{dataFrame from superFreq. such as data$allVariants$variants$variants$mySample}

\item{col}{colour. The colour of the lines. Default 'default' sets different colours on all lines.}

\item{lty}{linetype. The line type of the lines. Default 'default' sets different line types on all lines.}

\item{add}{logical. If the data should be plotted on top of whatever is already there. Default FALSE.}

\item{alpha}{numerical. A opaqueness multiplier. Default 1.}

\item{xlab}{character. The X label. Default "sample'}

\item{ylab}{character. The Y label. Default "clonality'}

\item{lwd}{linetype. The line width of the lines. Default 'default' sets different line width from accuracy.}

\item{errorBars}{logical. If error bars are used. Default TRUE.}

\item{setPar}{logical. If the margins are set. Default TRUE.}

\item{legend}{logical. If a legend is plotted. Default TRUE.}

\item{labels}{logical. If labels are plotted. Default TRUE.}

\item{xlim}{numerical. The X limits. Default "default' set it depending on number of samples and clones.}

\item{genome}{character. The genome the sample is aligned to. 'hg19', 'hg38' or 'mm10'}

\item{xSpread}{numerical. The stread of the points over a sample. Default 0.25.}

\item{sampleOrder}{character. The order of the samples. Default NA retains the order in the input.}

\item{...}{remaining arguments are passed to base plot(...) if add=F, otherwise ignored.}
}
\description{
plots line plots of clonalities
}
\details{
This function plots the story lines from a superFreq analysis.
         Can be used both for individual mutations or for clones.
}
