% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVEP.R
\name{postAnalyseVEP}
\alias{postAnalyseVEP}
\title{Runs VEP on a R and plot directory that superFreq has been run on.}
\usage{
postAnalyseVEP(outputDirectories, inputFiles = NA, metaData = NA,
  genome = "hg19", cpus = 1, cosmicDirectory = "", forceRedo = F)
}
\arguments{
\item{outputDirectories}{named list with Rdirectory and plotDirectory.}

\item{inputFiles}{named list as in analyse.}

\item{genome}{character: 'hg19' or 'mm10'}

\item{cpus}{integer. maximum number of parallel threads.}

\item{cosmicDirectory}{character: path to COSMIC directory. (optional)}

\item{forceRedo}{boolean: if TRUE, previously saved data is ignored and overwritten.}
}
\description{
Runs VEP on a R and plot directory that superFreq has been run on.
}
\details{
Runs VEP on the variants, and saves the variants (retrivable with loadData) with the addition
         information from VEP, and COSMIC if provided.
}

