% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateHTML.R
\name{printHTML}
\alias{printHTML}
\title{generates an HTML to navigate superFreq output}
\usage{
printHTML(metaDataFile, outputFile)
}
\arguments{
\item{metaDataFile}{Character. A path to the meta data file used for the run.}

\item{outputFile}{Character. A path to where the HTML is stored. Has to be inside the plotDirectory for now.}
}
\description{
generates an HTML to navigate superFreq output
}
\details{
This function is in pre-alpha development as of version 0.9.23. Feel free (please!) to add to this on github even without any html experience. I dont have any. The function generates a .html that should link the plots and explanation of how to interpret the plots, how they were generated and how to get the raw data for further analysis. Ideally there should be a good landing page as well for overview.
}
\examples{
\dontrun{
metaDataFile = 'metaData.tsv'
plotDirectory='plots'

printHTML(metaDataFile=metaDataFile, outputFile=paste0(plotDirectory, '/superFreq.html'))
}
}
