% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVariantAnnotation.R
\name{qToGRanges}
\alias{qToGRanges}
\title{Transforms a variant q data frame into granges}
\usage{
qToGRanges(q, genome = "hg19", addChrToSeqnames = F, seqlengths)
}
\description{
Transforms a variant q data frame into granges
}
