% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{requireFileExists}
\alias{requireFileExists}
\title{checks if a file exists, and creates an error if it doesnt.}
\usage{
requireFileExists(file)
}
\description{
checks if a file exists, and creates an error if it doesnt.
}
\details{
Internal method to give approapriate errors for missing files.
}
\examples{
requireFileExists('a/path/to/a/file.txt')
}

