% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEexon.R
\name{runDE}
\alias{runDE}
\title{Run differential coverage analysis.}
\usage{
runDE(bamFiles, names, externalNormalBams, captureRegions, Rdirectory,
  plotDirectory, normalRdirectory, settings = list(), genome = "hg19",
  cpus = 1, mode = "exome", forceRedoFit = F, forceRedoCount = F,
  forceRedoNormalCount = F)
}
\arguments{
\item{bamFiles}{vector of paths to the bamfiles to be analysed}

\item{names}{vector of names of the samples associated with the bam files.}

\item{externalNormalBams}{vector of paths to the pool of normal samples.}

\item{captureRegions}{GRanges capture regions, as you get from importCaptureRegions.}

\item{Rdirectory}{The directory to save or load the output to/from.}

\item{plotDirectory}{The directory to plot results and diagnosis to.}

\item{normalRdirectory}{The directory to save or load the pool of normal counts to/from.}

\item{settings}{Settings as passed to analyse(...).}

\item{genome}{Character string of the genome, such as 'hg19', 'hg38' or 'mm10'. Defaults to 'hg19'.}

\item{cpus}{The number of parallel cpus to be used. Defaults to 1.}

\item{forceRedoFit}{Boolean, forcing the differential analysis to be redone even if saved data is
present. Defaults to FALSE.}

\item{forceRedoCount}{Boolean, forcing the counting to be redone even if saved data is
present. Defaults to FALSE.}

\item{forceRedoNormalCount}{Boolean, forcing the counting of normal samples to be redone even
if saved data is present. Defaults to FALSE.}
}
\description{
Run differential coverage analysis.
}
\details{
Reads in the bed file into GRanges format, with extra columns for GC and binding strength.
}
