% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortAnalyseBatch.R
\name{runSummaryPostAnalysis}
\alias{runSummaryPostAnalysis}
\title{runs analyses across all samples in the batch}
\usage{
runSummaryPostAnalysis(
  Rdirectory,
  plotDirectory,
  metaDataFile,
  genome,
  resourceDirectory = "superFreqResources",
  GoI = NA,
  GoIakas = c(),
  excludeIndividuals = c(),
  excludeSamples = c(),
  cpus = 1
)
}
\arguments{
\item{Rdirectory}{character: path to the Rdirectory of the batch run.}

\item{plotDirectory}{character: path to the plotDirectory of the batch run.}

\item{metaDataFile}{character: path to the tab separated meta data file.}

\item{genome}{character: The genome build. 'hg19', hg38' or 'mm10'. Default 'hg19'}

\item{GoI}{character vector: The genes of interest to be highlighted in the analysis. Default NA uses genes that are mutated in the cohort that are also in COSMIC census or pathogenic in ClinVar.}

\item{GoIakas}{named character vector: changes the displayed name of the GoI. For example c('BCL2L11'='BIM', 'BCL2L1'='BCLxL') to display the commonly used gene names BIM and BCLxL instead of the SYMBOL names BCL2L11 and BCL2L1 used by superFreq. Default c().}

\item{excludeIndividuals}{character vector: Exclude these individuals from the analysis. Default c().}

\item{excludeSamples}{character vector: Exclude these samples from the analysis. Default c().}

\item{cpus}{integer: maximum number of threads used. default 1.}
}
\description{
runs analyses across all samples in the batch
}
\details{
This function carries out a number analyses across the individuals analysed in the top level Rdirectory. The output is placed in the plotDirectory, in a subdirectory called cohortWide. Probably convenient to not name individuals "cohortWide", or that individual will have to share directory with these plots.
}
\examples{
\dontrun{
superFreq(metaDataFile='metaData.tsv',
          Rdirectory='R',
          plotDirectory='plots',
          cpus=4,
          genome='hg38')

runSummaryPostAnalysis(metaDataFile='metaData.tsv',
          Rdirectory='R',
          plotDirectory='plots',
          cpus=4,
          genome='hg38')

}
}
