% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVEP.R
\name{runVEP}
\alias{runVEP}
\title{Run VEP on the provided variants}
\usage{
runVEP(variants, plotDir, cpus = 1, genome = "hg19", vepCall = "vep",
  forceRedoVEP = F, rareGermline = T)
}
\arguments{
\item{variants}{variants: The variants to be VEPed.}

\item{plotDir}{character: The plot directory where outputSomatics have been run.}

\item{cpus}{integer: The number of cpus to be used as most.}

\item{forceRedoVEP}{Logical: if VEP should be rerun even if saved data already exists.}
}
\description{
Run VEP on the provided variants
}
\details{
This function calls VEP on the output from outputSomaticVariants. For this, VEP needs to be callable by system('vep').
}
