% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVEP.R
\name{severityToType}
\alias{severityToType}
\title{The variant effect as function of severity}
\usage{
severityToType(severity)
}
\arguments{
\item{severity}{integer: The severity rank.}
}
\description{
The variant effect as function of severity
}
\details{
Translated back from severity rank to the variant effect.
}

