% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{shrug}
\alias{shrug}
\title{shruggie}
\usage{
shrug()
}
\description{
shruggie
}
\details{
¯\_(ツ)_/¯
}
\examples{
cat("Sample quality is too low to tell if your favourite gene is mutated or not.", shrug(), '\n')
}
