% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{superFreq}
\alias{superFreq}
\title{Wrapper to run default superFreq analysis}
\usage{
superFreq(metaDataFile, captureRegions, normalDirectory, Rdirectory,
  plotDirectory, reference, genome = "hg19", BQoffset = 33, cpus = 3,
  outputToTerminalAsWell = T, forceRedo = forceRedoNothing(),
  normalCoverageDirectory = "", systematicVariance = 0.02, maxCov = 150,
  cloneDistanceCut = -qnorm(0.01), dbSNPdirectory = "superFreqDbSNP",
  cosmicDirectory = "superFreqCOSMIC", mode = "exome", splitRun = F,
  participants = "all", manualStoryMerge = F, vepCall = "vep")
}
\arguments{
\item{metaDataFile}{Character. A path to a tab separated files with headers:
BAM: path (absolute or relative from metaData file) to bam file of sample.
     An index file is required as well.\cr
VCF: path (absolute or relative from metaData file) to vcf file of sample.
     This should include both somatic and germline variants. The variants
     will undergo further filtering, so it is not a problem if the .vcf
     includes false positives, although large number of entries in the
     .vcf increases run time.\cr
NAME: unique identifier of the sample. The names will be converted to
      standard R names, which involves replacing separators (such as space,
      dash or underscore) with dots.\cr
INDIVIDUAL: unique identifier of the indiviudal the sample comes from.
            This information is used pair up matched normal samples for
            somatic mutation filtering, and for identifying germline
            heterozygous SNPs. This column is also used to group samples
            that should be compared to each other: all samples from the
            same individual are analysed together, and mutations are
            tracked between the samples.\cr
NORMAL: Should be YES if the sample is normal, NO otherwise. A normal
        sample is assumed to have no somatic mutations. In practice, a
        tumor burden below a few percent works effectively as a normal,
        but tumor burdens above 5% in a sample marked as normal can cause
        severely reduced sensitivity in cancer samples from the same
        indiviudal.\cr
TIMEPOINT: Mostly used as label in plots together with the sample. Can
           be Diagnosis, Relapse, resistant or similar labels. Does not
           influence the analyss otherwise.}

\item{captureRegions}{Character. A path to a bed file of the unpadded capture regions.}

\item{normalDirectory}{Character. A path to a directory containing reference normals.
The directory should contain (links to) indexed .bam for the pool of
normal samples. These samples can, but dont have to, be matched normals
of the analysed samples. They need to be from the same capture as the
analysed samples, and preferably sequenced in the same lab. At least
two are required, but more are better (but slower). 10 is a good number.}

\item{Rdirectory}{Character. A path to the directory where data will be stored.
The directory will be created if needed, but the parent directory must exist.}

\item{plotDirectory}{Character. A path to the directory where plots and output is placed.
The directory will be created if needed, but the parent directory must exist.}

\item{reference}{Character. Path to the indexed fasta file the bams were aligned to.}

\item{genome}{Character. The reference genome the samples are aligned to, such as "hg19", "hg38"
or "mm10". Defaults to "hg19".}

\item{BQoffset}{Integer. The offset of the base quality encoding. Defaults to 33.}

\item{cpus}{Integer. Maximum number of parallel threads. Often threaded over chromsomes,
so numbers like 4, 6, 8, 12, 24 are good for human genomes. Default 3.}

\item{outputToTerminalAsWell}{Boolean. If the rather verbose log output should be sent to
terminal as well as to the logfile (in the Rdirectory).
Default TRUE.}

\item{forceRedo}{named list of Booleans. Controls which step are retrieved from previously
                 saved data, and which are redone even if previous results are saved.
                 If a step is forced to be redone, depending downstream steps are also
                 redone. The named entries in the list should be:\cr

                 forceRedoCount: The read counting over capture regions for the samples.\cr
                 forceRedoNormalCount: The read counting over capture regions for the
                                       set of reference normals.\cr
                 forceRedoFit: The differential coverage analysis.\cr
                 forceRedoVolcanoes: Volcano plots of the differential coverage.\cr
                 forceRedoDifferentRegions: Top table output of gained/lost genes/exons.\cr
                 forceRedoSNPs: Deprecated, usually not called.\cr
                 forceRedoVariants: Import and QC of the variants pointed to in the VCF. \cr
                 forceRedoNormalSNPs: Deprecated, usually not called.\cr
                 forceRedoNormalVariants: Import and QC of the variants in the set of
                                          reference normals.\cr
                 forceRedoMatchFlag: QC of variants using the set of reference normals.
                                     Somatic SNV are also separated from germline.\cr
                 forceRedoScatters: Scatter plots of the SNV frequencies between samples.\cr
                 forceRedoOutputSomatic: Top tables of somatic SNVs by sample.\cr
                 forceRedoNewVariants: Top table of new somatic SNVs compared to other
                                       samples.\cr
                 forceRedoSNPprogression: Heatmap of SNV frequencies over multiple samples.\cr
                 forceRedoCNV: Copy number calling.\cr
                 forceRedoCNVplots: Copy number plots.\cr
                 forceRedoSummary: Summary plot of SNVs and CNVs over all samples.\cr
                 forceRedoStories: Clonal tracking of somatic mutations.\cr
                 forceRedoRiver: Plots of the clonal evolution.\cr
                 forceRedoVEP: VEP call (Variant Effect Predictor) and COSMIC linking.

                 The default value is forceRedoNothing(), setting all entries to FALSE.
                 To redo everything, set to forceRedoEverything().}

\item{systematicVariance}{Numeric. A measure of the expected systematic variance in coverage
log fold change. Larger value decreases false CNV calls, smaller
values increase sensitivity. Defaults to 0.03.}

\item{maxCov}{Integer. The coverage at which systematic variance of SNV frequencies are equal
to the Poissonian variance. Larger values increase sensitivity in CNV
calls, smaller values decrease false positives. Default 150.}

\item{dbSNPdirectory}{Character. The location of the directory where the superFreq dbSNP data
is located. If the directory doesn't exist, it will be created and the
data will be downloaded. Defaults to superFreqDbSNP (in the directory
where R is run).}

\item{cosmicDirectory}{Character. The location of the directory where the superFreq COSMIC data
is located. If the directory doesn't exist, it will be created and the
data will be downloaded. Defaults to superFreqCOSMIC (in the directory
where R is run).}

\item{mode}{Character. The mode to run in. The default 'Exome' is almost always used. If running on RNA,
switch to "RNA", which seems to work decently, but beware of limitations of the data.}

\item{splitRun}{logical. If the superFreq run should be split by individual. Runs each individual separately,
which decrease memory footprint. Default is FALSE, but TRUE is recommended when running on
multiple individuals in the same batch.}

\item{participants}{character. When run in split mode, this allows you to run on a subset of individuals.
Default is "all" which runs on all individuals.}

\item{manualStoryMerge}{logical. A flag which gives you some manual control over the merging of mutations into clones.
Requires babysitting and not recommended. Do not press this button. Default FALSE.}

\item{vepCall}{character. The call for the variant effect predictor on the command line. Default "vep",
otherwise try the path to variant_effect_predictor.pl.}
}
\description{
Wrapper to run default superFreq analysis
}
\details{
This function runs a full SNV, CNV and clonality analysis of the input exome data. Read more on the manual on https://github.com/ChristofferFlensburg/superFreq/blob/master/manual.pdf
}
\examples{
\dontrun{
#minimal example to get it running, assuming data fits the defaults.
#you really want to skim through the other settings before running though.
metaDataFile = 'metaData.tsv'
captureRegions='captureRegions.bed'
normalDirectory = '../referenceNormals'
reference = '../reference/hg19.fa'
Rdirectory = 'R'
plotDirectory='plots'

data = superFreq(metaDataFile, captureRegions, normalDirectory,
                 Rdirectory, plotDirectory, reference)
}
}

