% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{superInputFiles}
\alias{superInputFiles}
\title{sets up and checks the input files for superFreq}
\usage{
superInputFiles(metaData = "", captureRegions = "", normalDirectory = "",
  dbSNPdirectory = "", reference = "", normalCoverageDirectory = "")
}
\description{
sets up and checks the input files for superFreq
}
\details{
Returns the input files in a format to be passed to analyse(). Returns detailed information about the required input files if missing.
}
\examples{
superInputFiles()
}

