% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{superOutputDirectories}
\alias{superOutputDirectories}
\title{Sets up and checks the output directories.}
\usage{
superOutputDirectories(Rdirectory = "", plotDirectory = "")
}
\description{
Sets up and checks the output directories.
}
\details{
Outputs some infomration about what the paths should be if called empty.
}
\examples{
superOutputDirectories()
}

