% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{superVersion}
\alias{superVersion}
\title{Return the current version of superFreq.}
\usage{
superVersion()
}
\description{
Return the current version of superFreq.
}
\details{
this is the version that goes in the logfile.
         First digit means very large changes.
         Second digit means algorithm changes that affects output.
         Third digit is debugging that shouldn't affect output of completed runs.
         1.0.0 will be the version used in the performance testing in the first preprint.
}

