% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVEP.R
\name{typeToSeverity}
\alias{typeToSeverity}
\title{Ranks the variant effects}
\usage{
typeToSeverity(type, annotationMethod = "VEP")
}
\arguments{
\item{type}{character: The effect.}
}
\description{
Ranks the variant effects
}
\details{
Returns a low score for severe effects, higher for less severe, 100 for unknown. 11 and below alters the protein.
}
