% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVariantAnnotation.R
\name{VAconsequenceToSeverityRank}
\alias{VAconsequenceToSeverityRank}
\title{coding' is used by the location run to cover all the coding variants, and will be replaced}
\usage{
VAconsequenceToSeverityRank(consequence)
}
\description{
coding' is used by the location run to cover all the coding variants, and will be replaced
}
