% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runVariantAnnotation.R
\name{VAseverityToConsequence}
\alias{VAseverityToConsequence}
\title{coding' is used by the location run to cover all the coding variants, and will be replaced}
\usage{
VAseverityToConsequence(severity)
}
\description{
coding' is used by the location run to cover all the coding variants, and will be replaced
}
