% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputSomaticVariants.R
\name{writeToVCF}
\alias{writeToVCF}
\title{exports variants to VCF}
\usage{
writeToVCF(q, vcfFile, genome = "hg19", snvOnly = F)
}
\arguments{
\item{q}{A variant data frame from superFreq.}

\item{vcfFile}{The path to the output file.}

\item{genome}{The genome, such as 'hg19', 'hg38' or 'mm10'. Defaults to 'hg19'.}

\item{SNVonly}{boolean. Set to TRUE to only output SNVs, not indels. Defaults to FALSE.}
}
\description{
exports variants to VCF
}
\details{
This function outputs superFreq variants to a VCF for access from other software.
}
\examples{
\dontrun{
data = loadData(Rdirectory)
q = data$allVariants$variants$variants$mySample
writeToVCF(q, 'mySample.superFreq.vcf')
}
}
