% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFitPlots.R
\name{xToChr}
\alias{xToChr}
\title{Extract chromsome of genomic coordinate}
\usage{
xToChr(x, genome = "hg19")
}
\arguments{
\item{x}{integer: The genomic coordinate.}

\item{genome}{character: the genome. default "hg19"}
}
\description{
Extract chromsome of genomic coordinate
}
\details{
Extracts which chromsome a genomic coordinate is on.
}
\examples{
xToChr(1e9)
xToChr(1e9, genome='mm10')
}
