% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importCaptureRegions.R
\name{xToGene}
\alias{xToGene}
\title{find the gene closest to the genomic coordinate x}
\usage{
xToGene(x, genome = "hg19", saveDirectory = "", verbose = T)
}
\description{
find the gene closest to the genomic coordinate x
}

