% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFitPlots.R
\name{xToPos}
\alias{xToPos}
\title{Extract position of genomic coordinate}
\usage{
xToPos(x, genome = "hg19")
}
\arguments{
\item{x}{integer: The genomic coordinate.}

\item{genome}{character: the genome. default "hg19"}
}
\description{
Extract position of genomic coordinate
}
\details{
Extracts the position on the chromsome a genomic coordinate is on.
}
\examples{
xToPos(1e9)
xToPos(1e9, genome='mm10')
}
