% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDEexon.R
\name{plotColourScatter}
\alias{plotColourScatter}
\title{A better version of plot}
\usage{
plotColourScatter(x, y, xlab = "", ylab = "", col = "defaultBlue",
  main = "cor", add = F, cex = 1, showDensity = T, ...)
}
\arguments{
\item{x}{Numeric. The x coordinates.}

\item{y}{Numeric. The y coordinates.}

\item{add}{boolean. If adding the plot onto whatever is alrady there. Essentially turns plot() it into points().}

\item{showDensity}{Boolean. Adds a gradient through cyan and green in dense regions. Default TRUE.}

\item{...}{Remaining parameters are passed to plot(...), or to points(...) if add.}
}
\description{
A better version of plot
}
\details{
This plotting function does the same as plot() in essence, but the default looks nicer than the black rings, and overplotting smoothly transitions into heatmap representation. The price is that pdfs are a bit larger than from a plot() call.
}
\examples{
x = rnorm(10000, ((1:10000)/10000)^2, 0.01)
y = rnorm(10000, ((1:10000)/10000)^3, 0.01)
plotColourScatter(x, y)

}
